# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farming Service API",
        "description": "API documentation for Farming Service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Add a new service",
                "description": "Create a new service with name, description, price, and enabled status.",
                "operationId": "b05b71c30267faa468172458580fb6fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_name",
                                    "price"
                                ],
                                "properties": {
                                    "service_name": {
                                        "type": "string",
                                        "example": "Lawn Mowing"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A basic lawn mowing service"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "is_enabled": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/201"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Display a specific service",
                "description": "Get the details of a specific service.",
                "operationId": "42775b582b563bc36a9e06ae850100bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Services"
                ],
                "summary": "Update a service",
                "description": "Update the details of an existing service.",
                "operationId": "38f6a6cc2f5cd3dfdf39805e7fa21996",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_name",
                                    "price"
                                ],
                                "properties": {
                                    "service_name": {
                                        "type": "string",
                                        "example": "Lawn Mowing"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description of service"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 60
                                    },
                                    "is_enabled": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Services"
                ],
                "summary": "Delete a service",
                "description": "Remove a service from the database.",
                "operationId": "c189d6ab2d037e8fb4228a535ae6a399",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "responses": {
...
            "500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Something went wrong!"
                                },
                                "error": {
                                    "type": "string",
                                    "example": "Internal Server Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "200": {
                "description": "Successful operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Operation successful"
                                },
                                "data": {
                                    "type": "object",
                                    "example": []
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "201": {
                "description": "Resource created successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Resource created successfully"
                                },
                                "data": {
                                    "type": "object",
                                    "example": []
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "Validation error - invalid input data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "error": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}